Crate spinoso_string

A String object holds and manipulates an arbitrary sequence of bytes, typically representing characters.

Structs

Bytes

Immutable String byte iterator.

Center

An iterator that yields a byte string centered within a padding byte string.

Chars

Codepoints

An iterator that yields a u32 codepoints from a String.

Inspect

An iterator that yields a debug representation of a String and its byte contents as a sequence of chars.

IntoIter

An iterator that moves out of a string.

InvalidCodepointError

InvalidEncodingError

Error returned when failing to deserialize an Encoding.

Iter

Immutable String byte slice iterator.

IterMut

Mutable String byte iterator.

RawParts

A wrapper around the decomposed parts of a Vec<T>.

String

Enums

CaseFold

Unicode case folding strategies.

CenterError

Error returned when failing to construct a Center iterator.

CodepointsError

Error returned when failing to construct a Codepoints iterator/

Encoding

An Encoding instance represents a character encoding usable in Ruby.

OrdError